requirements.txt
# 论文推送工作流依赖
# pip install -r requirements.txt

requests>=2.31.0        # HTTP请求
arxiv>=2.1.0            # arXiv API 客户端
openai>=1.12.0          # DeepSeek兼容OpenAI SDK
pyyaml>=6.0.1           # YAML配置解析
jinja2>=3.1.3           # HTML模板引擎
schedule>=1.2.1         # 本地定时任务(可选)
python-dateutil>=2.8.2  # 日期处理
beautifulsoup4>=4.12.0  # HTML处理
lxml>=5.1.0             # XML解析
